Rewrite setup.sh to do more stuff
setup.sh was completely changed to do a bit more work than before. For
instance, we no longer assume that the user is using SSH or HTTPS, an
support both URL types.

setup.sh can also be used in the future for more complex tasks beyond
what can be put inside the .git/config-kind of file.